add some git and pair plugin

+++ install.sh
@@ -49,9 +49,20 @@ function main() {
 	parse_cmdline_arguments "$@"
 
 	echo "===== Ubuntu install ==="
-	sudo apt update
-	sudo apt install -y tmux wget
+	if [ ! -f /usr/bin/tmux ] || [ ! -f /usr/bin/wget ]; then
+		sudo apt update
+		sudo apt install -y tmux wget
+	fi
 
+	echo "===== Install Lazy Git ="
+	if [ ! -f /usr/local/bin/lazygit ]; then
+		local LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
+		curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
+		tar xf lazygit.tar.gz lazygit
+		sudo install lazygit /usr/local/bin
+		rm -f lazygit*
+	fi
+	
 	echo "===== Install nvim ====="
 	if [ ! -f /usr/bin/nvim ]; then
 		wget https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz -O nvim.tar.gz
